fix for finding list view in a different placeholder
It was possible for the findTargetListView function to short-circuit
before it found a legitimate list view target. This only happened in a
very specific condition:

When moving between 2 pages using the same layout, where the
ListViewControl widget was in a different placeholder than the List
View, the findTargetListView function would reach the incubator node
and give up before searching the children of that node. In this
particular case, the incubator has 2 children: one for each placeholder
being updated in the page navigation. So, it is legitimately possible
for the first common ancestor of a LVC widget and a LV to be the
incubator node itself.

As such, I have moved the if()...break statement below the for...return
loop.